requirements.txt
streamlit
python-dotenv
sentence-transformers
chromadb
pymupdf

langchain==0.0.353
langchain-groq
